refactor(command): split handlers by feature
