+++ src/command/handlers/help.rs
@@ -1,0 +1,26 @@
+use crate::error::AppResult;
+
+use super::super::core::{close_help as close_help_core, open_help as open_help_core};
+use super::super::dispatch::{CommandExecContext, CommandExecution};
+
+pub(in crate::command) fn open_help(
+    ctx: &mut CommandExecContext<'_>,
+) -> AppResult<CommandExecution> {
+    let result = open_help_core(ctx.app)?;
+    Ok(CommandExecution::from_notice_result(result))
+}
+
+pub(in crate::command) fn close_help(
+    ctx: &mut CommandExecContext<'_>,
+) -> AppResult<CommandExecution> {
+    let result = close_help_core(ctx.app)?;
+    Ok(CommandExecution::from_notice_result(result))
+}
+
+pub(in crate::command) fn help_scroll(
+    ctx: &mut CommandExecContext<'_>,
+    delta: isize,
+) -> AppResult<CommandExecution> {
+    ctx.app.scroll_help_by(delta);
+    Ok(CommandExecution::applied())
+}
