+++ src/command/handlers/mod.rs
@@ -1,0 +1,24 @@
+mod control;
+mod debug;
+mod help;
+mod history;
+mod layout;
+mod navigation;
+mod outline;
+mod palette;
+mod search;
+mod viewport;
+
+pub(super) use control::{cancel, quit};
+pub(super) use debug::{debug_status_hide, debug_status_show, debug_status_toggle};
+pub(super) use help::{close_help, help_scroll, open_help};
+pub(super) use history::{history_back, history_forward, history_goto, open_history};
+pub(super) use layout::{page_layout_single, page_layout_spread};
+pub(super) use navigation::{first_page, goto_page, last_page, next_page, prev_page};
+pub(super) use outline::{open_outline, outline_goto};
+pub(super) use palette::{close_palette, open_palette};
+pub(super) use search::{
+    next_search_hit, open_search, open_search_results, prev_search_hit, search_result_goto,
+    submit_search,
+};
+pub(super) use viewport::{pan, set_zoom, zoom_in, zoom_out, zoom_reset};
